[DEV-4856] component for share toolbar item
